cloud/requirements.txt
fastapi
uvicorn
torch
Pillow
numpy
gtts
